requirements.txt
streamlit>=1.32.0
anthropic>=0.40.0
python-dotenv>=1.0.0
duckduckgo-search>=7.0.0
google-genai>=1.0.0
